BB-941: Handle missing block keys gracefully.
We recalculate the entire course when they are found, so we know that
our aggregators reflect the current structure of the course.  This
should be a relatively rare occurrence.

This PR also fixes an issue where block aggregators were not being found
in the test compat library.  This made every test run act as though all
aggregators had been changed.

Finally, we add a test to cover the case where a block has been
completed but the block key is not included in the list of
changed_blocks, and so the aggregator is not updated.  A later run of the
aggregator will catch the new change.