fix: completion progress bar errors when user not enrolled in course
The CompletionDetailView raises an 404 if it is unable to calculation
the completion details for a specific user/course
combination (eg. the user isn't enrolled in the course).

CompletionProgressBarView calls this view to determine the percentage
to show to a learner.

A 404 in the main transaction, causes it to be rolled back, hence any
further db calls will break with the error.

> django.db.transaction.TransactionManagementError: An error occurred
in the current transaction. You can't execute queries until the end of
the 'atomic' block.

The change in this commit wraps the offending code in another
transaction so that the inner transaction is rolled back rather than
the main one.